Moved helper functions and classes
